+++ CHANGELOG.md
@@ -7,9 +7,13 @@ nav_order: 1
 
 # Changelog
 
+## [MAJOR.MINOR.PATCH] - YYYY-MM-DD
+
+- Add additional disk space support
+
 ## [3.5.1] - 2022-08-16
 
-- Add service disc space custom diff 404 error handling
+- Add service disk space custom diff 404 error handling
 - Fix VPC peering connection custom diff
 
 ## [3.5.0] - 2022-08-10